Introduce macro impl_numeric_bound_validator!
