+++ nutype_macros/src/float/models.rs
@@ -3,7 +3,7 @@ use proc_macro2::TokenStream;
 use std::fmt::Debug;
 
 use crate::common::models::{
-    Guard, NumericBoundValidator, RawGuard, SpannedItem, TypeTrait, TypedCustomFunction,
+    impl_numeric_bound_validator, Guard, RawGuard, SpannedItem, TypeTrait, TypedCustomFunction,
 };
 
 // Sanitizer
@@ -32,39 +32,7 @@ pub enum FloatValidator<T> {
     Finite,
 }
 
-impl<T: Clone> NumericBoundValidator<T> for FloatValidator<T> {
-    fn greater(&self) -> Option<T> {
-        if let FloatValidator::Greater(value) = self {
-            Some(value.clone())
-        } else {
-            None
-        }
-    }
-
-    fn greater_or_equal(&self) -> Option<T> {
-        if let FloatValidator::GreaterOrEqual(value) = self {
-            Some(value.clone())
-        } else {
-            None
-        }
-    }
-
-    fn less(&self) -> Option<T> {
-        if let FloatValidator::Less(value) = self {
-            Some(value.clone())
-        } else {
-            None
-        }
-    }
-
-    fn less_or_equal(&self) -> Option<T> {
-        if let FloatValidator::LessOrEqual(value) = self {
-            Some(value.clone())
-        } else {
-            None
-        }
-    }
-}
+impl_numeric_bound_validator!(FloatValidator);
 
 pub type SpannedFloatValidator<T> = SpannedItem<FloatValidator<T>>;
